+++ inc/lilyNotation/index.ts
@@ -1,0 +1,11 @@
+
+import {Note, Notation} from "./notation";
+import {fuzzyMatchNotations} from "./fuzzyMatch";
+
+
+
+export {
+	Note,
+	Notation,
+	fuzzyMatchNotations,
+};